hannah: audio: Increase the client capture read delay value
The write to pcm_data is waiting and is getting blocked.
The 2second keyword data is not being read fully in CPE mode
and is leading to second stage detection failure.
The detections are not happening as there are no extra
frames after the keyword is captured.

Increase the value of client capture read delay value increases
the value of buffer size so that the complete data can be captured
for the keyword to be detected properly in the CPE mode.

Change-Id: Ie1276b67cfb39e7619896c2dff60398feccd928d